requirements.txt
numpy>=1.26.0,<2.1.0
opencv-python-headless==4.8.0.76
streamlit==1.27.0
Pillow<10
geocoder
piexif
